(trunk, libT) #5709 'Fix an instance of time_t printing' -- fixed, patch by naddy

+++ libtransmission/peer-mgr.c
@@ -2892,9 +2892,9 @@ rechokeDownloads (tr_swarm * s)
         const double mult = MIN (timeSinceCancel, maxHistory) / (double) maxHistory;
         const int inc = maxIncrease * mult;
         maxPeers = s->maxPeers + inc;
-        tordbg (s, "time since last cancel is %li -- increasing the "
+        tordbg (s, "time since last cancel is %lli -- increasing the "
                    "number of peers we're interested in by %d",
-                   timeSinceCancel, inc);
+                   (long long)timeSinceCancel, inc);
       }
   }
